# docs/spotlight/purchase-flow.openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace — Flujo de Compra (Pedidos y Pagos)
  version: 1.0.0
  description: |
    Endpoints del flujo de compra del marketplace.
    Incluye creación de pedidos, listado, detalle, procesamiento de pagos
    y gestión de estados por parte del vendedor.

    ## Flujo de Compra
    1. **Comprador** selecciona productos y crea un pedido (`POST /orders`)
    2. **Comprador** procesa el pago (`POST /payments`)
    3. **Vendedor** marca como enviado (`PUT /seller/orders/{id}/status` → shipped)
    4. **Vendedor** confirma entrega (`PUT /seller/orders/{id}/status` → delivered)

    ## Estados del Pedido
    ```
    pending → paid → shipped → delivered
       ↓        ↓
    cancelled cancelled
    ```
servers:
  - url: http://localhost:8000/api
    description: Servidor de Desarrollo Local

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token generado al hacer login con Sanctum.

  schemas:
    # ── Request Schemas ─────────────────────────────────────
    CreateOrderRequest:
      type: object
      required:
        - items
      properties:
        notes:
          type: string
          nullable: true
          maxLength: 1000
          example: Entregar en la puerta principal
        items:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - product_id
              - quantity
            properties:
              product_id:
                type: integer
                example: 1
              quantity:
                type: integer
                minimum: 1
                example: 2

    UpdateOrderStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [paid, shipped, delivered, cancelled]
          example: shipped

    CreatePaymentRequest:
      type: object
      required:
        - order_id
        - method
      properties:
        order_id:
          type: integer
          example: 1
        method:
          type: string
          enum: [credit_card, paypal, transfer, cash]
          example: credit_card

    # ── Response Schemas ────────────────────────────────────
    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        product_id:
          type: integer
          example: 5
        quantity:
          type: integer
          example: 2
        unit_price:
          type: string
          example: "150.00"
        subtotal:
          type: number
          format: float
          example: 300.00
        product:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            price:
              type: string

    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        buyer_id:
          type: integer
          example: 3
        status:
          type: string
          enum: [pending, paid, shipped, delivered, cancelled]
          example: pending
        total:
          type: string
          example: "350.00"
        notes:
          type: string
          nullable: true
          example: Entregar rápido
        buyer:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            email:
              type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        payment:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            method:
              type: string
            status:
              type: string
            amount:
              type: string
            transaction_id:
              type: string
            paid_at:
              type: string
              format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    PaymentResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        order_id:
          type: integer
          example: 1
        method:
          type: string
          enum: [credit_card, paypal, transfer, cash]
          example: credit_card
        status:
          type: string
          enum: [pending, completed, failed, refunded]
          example: completed
        amount:
          type: string
          example: "350.00"
        transaction_id:
          type: string
          example: TXN-A1B2C3D4E5F6G7H8
        paid_at:
          type: string
          format: date-time
          nullable: true
        order:
          type: object
          properties:
            id:
              type: integer
            status:
              type: string
            total:
              type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object

    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: object
          nullable: true

    PaginatedMeta:
      type: object
      properties:
        current_page:
          type: integer
        last_page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer

paths:
  # ── Pedidos (Comprador) ─────────────────────────────────────
  /orders:
    post:
      summary: Crear pedido
      description: |
        Permite a un **comprador** crear un nuevo pedido con uno o más productos.
        - Valida que cada producto exista y tenga stock suficiente.
        - Descuenta automáticamente el stock de cada producto.
        - Calcula el total sumando `unit_price × quantity` de cada item.
      tags:
        - Pedidos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Pedido creado exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderResponse'
        '403':
          description: Solo compradores pueden crear pedidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Error de validación o stock insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

    get:
      summary: Listar mis pedidos (comprador)
      description: |
        Devuelve los pedidos del **comprador autenticado** con paginación.
        Se puede filtrar por estado usando el query parameter `status`.
      tags:
        - Pedidos
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, paid, shipped, delivered, cancelled]
          description: Filtrar por estado del pedido
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Lista de pedidos del comprador
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderResponse'
                  meta:
                    $ref: '#/components/schemas/PaginatedMeta'

  /orders/{id}:
    get:
      summary: Detalle de pedido
      description: |
        Muestra el detalle completo de un pedido.
        - **Comprador**: solo ve sus propios pedidos.
        - **Vendedor**: solo si tiene productos en el pedido.
        - **Admin**: ve cualquier pedido.
      tags:
        - Pedidos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalle del pedido
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderResponse'
        '403':
          description: No autorizado para ver este pedido
        '404':
          description: Pedido no encontrado

  # ── Pedidos (Vendedor) ──────────────────────────────────────
  /seller/orders:
    get:
      summary: Pedidos recibidos (vendedor)
      description: |
        Lista los pedidos que contienen al menos un producto del **vendedor autenticado**.
        Filtrable por estado.
      tags:
        - Pedidos
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Pedidos del vendedor
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderResponse'
                  meta:
                    $ref: '#/components/schemas/PaginatedMeta'
        '403':
          description: Solo vendedores pueden acceder

  /seller/orders/{id}/status:
    put:
      summary: Actualizar estado del pedido (vendedor)
      description: |
        Permite al **vendedor** actualizar el estado de un pedido que contenga sus productos.

        ### Transiciones válidas
        | Desde     | Hacia                |
        |-----------|----------------------|
        | pending   | paid, cancelled      |
        | paid      | shipped, cancelled   |
        | shipped   | delivered            |
        | delivered | (ninguna)            |
        | cancelled | (ninguna)            |

        Al cancelar un pedido, se restaura el stock de los productos.
      tags:
        - Pedidos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '200':
          description: Estado actualizado exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderResponse'
        '403':
          description: No tienes productos en este pedido
        '422':
          description: Transición de estado no válida

  # ── Pagos ───────────────────────────────────────────────────
  /payments:
    post:
      summary: Procesar pago de un pedido
      description: |
        Crea un registro de pago asociado a un pedido **pendiente**.
        - Solo el **comprador** dueño del pedido puede pagar.
        - Simula el procesamiento y genera un `transaction_id` único.
        - Actualiza automáticamente el estado del pedido a `paid`.
      tags:
        - Pagos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Pago procesado exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaymentResponse'
        '403':
          description: No puedes pagar un pedido que no es tuyo
        '422':
          description: Pedido ya procesado o ya tiene un pago asociado

  /payments/{id}:
    get:
      summary: Estado del pago
      description: |
        Consulta el estado y detalle de un pago específico.
        - **Comprador**: solo sus propios pagos.
        - **Vendedor**: solo pagos de pedidos con sus productos.
        - **Admin**: cualquier pago.
      tags:
        - Pagos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalle del pago
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaymentResponse'
        '403':
          description: No autorizado
        '404':
          description: Pago no encontrado